Variant Overhaul to Flags
Getting rid of the variant argument and replacing it with flags for button, border, and fill stylings.

Also removes any of the draw arc code and instead repeats tick or dot draws. This helps with some of the weird alpha overlaps in the middle of tracks sometimes.

Lastly, supports for new styles included rounded tracks and wiper styles that fill from either end or the middle of the control region of the knob